preparation to add more attribute to csv file

+++ solver.py
@@ -1,7 +1,6 @@
 import pathlib
 import networkx as nx
 import tempfile
-import os
 import sys
 import subprocess
 import codecs
@@ -12,7 +11,6 @@ from datetime import datetime
 
 
 def read_gr(file):
-    sys.stdout.write(str(file) + '\n')
     tmp = tempfile.NamedTemporaryFile()
     description = []
     # remove comments save graph description
@@ -30,10 +28,9 @@ def read_gr(file):
     if len(description) == 1:
         info = codecs.decode(tmp.readline().strip(), "UTF-8").split(" ")
         if len(info) > 4:
-            sys.stderr.write("to many arguments\n")
+            sys.stderr.write("to many arguments: {}\n".format(str(file)))
             return
         if info[0] == 'p' and info[2].isnumeric() and info[3].isnumeric():
-
             r_vertices = int(info[2])
             r_edges = int(info[3])
             # read remaining edges and build graph
@@ -42,7 +39,8 @@ def read_gr(file):
                 if len(edge) == 2 and edge[0].isnumeric() and edge[1].isnumeric():
                     g.add_edge(edge[0], edge[1])
                 else:
-                    sys.stderr.write("expected two number parameter!\n")
+                    sys.stderr.write(
+                        "expected two number parameter: {}\n".format(str(file)))
                     return
 
             b_vertices = g.number_of_nodes()
@@ -58,21 +56,25 @@ def read_gr(file):
 
                 tmp.seek(0)
                 to_sif(file, tmp)
+
+                # return initial graph to caller
+                return g
             else:
-                sys.stderr.write("wrong input format!\n")
+                sys.stderr.write("wrong input format: {}\n".format(str(file)))
                 return
 
         else:
-            sys.stderr.write("expected meta information on first line!\n")
+            sys.stderr.write(
+                "expected meta information on first line: {}\n".format(str(file)))
             return
     else:
-        sys.stderr.write("expected exactly one line with\n")
+        sys.stderr.write(
+            "expected exactly one line with: {}\n".format(str(file)))
         return
 
 
 def read_graph6(file):
-    sys.stdout.write(str(file) + '\n')
-    tmp = tempfile.TemporaryFile()
+    tmp = tempfile.NamedTemporaryFile()
     content = file.read_text()
 
     for i in iter(content.splitlines()):
@@ -82,18 +84,29 @@ def read_graph6(file):
         for (u, v) in g.edges(data=False):
             tmp.write(bytes("{:d} {:d}\n".format(u + 1, v + 1), "UTF-8"))
     tmp.seek(0)
+
+    g = nx.read_edgelist(tmp.name)
+    tmp.seek(0)
     to_sif(file, tmp)
+
+    # return initial graph to caller
+    return g
 
 
 def to_sif(file, tmp):
     name = file.name.split(".")[0]
     path = pathlib.Path(file).parent.absolute()
-    sif = path.joinpath(name + '.sif')
+    solution = path.joinpath(name)
 
-    with sif.open("w+") as f:
-        for i in tmp:
-            values = codecs.decode(i, "UTF-8").rstrip().split(" ")
-            f.write("{:d} xx {:d}\n".format(int(values[0]), int(values[1])))
+    # dont create .sif file if solution file already exists
+    if(not solution.exists()):
+        sif = path.joinpath(name + '.sif')
+
+        with sif.open("w+") as f:
+            for i in tmp:
+                values = codecs.decode(i, "UTF-8").rstrip().split(" ")
+                f.write("{:d} xx {:d}\n".format(
+                    int(values[0]), int(values[1])))
 
 
 def gml_to_list(path):
@@ -109,39 +122,61 @@ def gml_to_list(path):
         i.unlink()
 
 
-def main():
-    path = pathlib.Path.cwd().joinpath('test')
+def get_characteristics(graph):
+    return graph.size()
 
-    # convert to sif
-    for i in path.glob('**/*.*'):
-        if i.suffixes[0] == ".gr":
-            read_gr(i)
-        elif i.suffixes[0] == ".graph6":
-            read_graph6(i)
-        elif i.suffixes[0] == ".td":
-            i.unlink()
+
+def main():
 
     # check for yoshiko
     solver = pathlib.Path.cwd().joinpath('yoshiko')
-    if not os.path.isfile(solver):
-        sys.exit("Error: binary not found: " + str(solver))
+    if not solver.is_file():
+        sys.exit("Err: binary not found: " + str(solver))
 
-    # solve and store solution complexity
+    path = pathlib.Path.cwd().joinpath('test')
+    graphs = {}
+
+    # convert to .sif
+    for i in path.glob('**/*.*'):
+        if i.suffix == ".gr":
+            g = read_gr(i)
+            if isinstance(g, nx.classes.graph.Graph):
+                graphs.update({str(i.with_suffix('')): g})
+
+        elif i.suffix == ".graph6":
+            g = read_graph6(i)
+            if isinstance(g, nx.classes.graph.Graph):
+                graphs.update({str(i.with_suffix('')): g})
+
+        elif i.suffix == ".td":
+            i.unlink()
+
+    # solve and store important characteristics
     optima = pathlib.Path.cwd().joinpath('test', 'optimum.csv')
-    with open(optima, 'w') as f:
-        header = ['filename', 'seconds', 'complexity']
+    with open(optima, 'a+') as f:
+        header = ['filename', 'n', 'k']
         writer = csv.DictWriter(f, fieldnames=header)
-        writer.writeheader()
+
+        if (optima.stat().st_size == 0):
+            writer.writeheader()
 
         cpu = multiprocessing.cpu_count() - 2
+
         for i in path.glob('**/*.sif'):
             # solver cant solve empty files
-            if os.stat(i).st_size == 0:
+            if i.stat().st_size == 0:
                 i.unlink()
             else:
+                sys.stdout.write("solving: {}\n".format(str(i)))
                 dir = pathlib.Path(i).parent.absolute()
                 name = i.name.split(".")[0]
-                gml = dir.joinpath(name)
+
+                # as String: Key to dict "graphs"
+                # as Pathlib.path: Path to yoshiko solution
+                solution = dir.joinpath(name)
+                # path to solution file
+                gml = solution.with_suffix('.gml')
+
                 proc = subprocess.Popen([solver,
                                          "-f",
                                          i,
@@ -154,7 +189,7 @@ def main():
                                          "-threads",
                                          str(cpu),
                                          "-o",
-                                         gml],
+                                         solution],
                                         stdout=subprocess.PIPE,
                                         stdin=subprocess.PIPE)
 
@@ -163,17 +198,31 @@ def main():
                     outs, errs = proc.communicate(timeout=1800)
                     end = time.time()
 
+                    # "solution" points now to graph in dict
+                    # read into networkx and compute for csv:
+                    # count of vertices + edges
+                    # count of clusters
+                    graph = graphs.get(str(solution))
+                    ig = get_characteristics(graph)
+
+                    # "gml" points to solution file
+                    # read into networkx and compute for csv:
+                    # count of vertices + edges
+                    # count of clusters
+                    graph = nx.read_gml(gml)
+                    og = get_characteristics(graph)
+
                     complexity = codecs.decode(outs, 'UTF-8')
                     elapsed = end - start
                     writer.writerow({'filename': name.strip(),
-                                     'seconds': elapsed,
-                                     'complexity': complexity.strip()})
+                                     'n': elapsed,
+                                     'k': complexity.strip()})
                     i.unlink()
                 except subprocess.TimeoutExpired:
                     proc.kill()
                     writer.writerow({'filename': name.strip(),
-                                     'seconds': "Err",
-                                     'complexity': "Err"})
+                                     'n': "Err",
+                                     'k': "Err"})
 
     gml_to_list(path)
